🌟 add frame insights

+++ src/main/kotlin/io/github/theapache64/perfsheet/model/Node.kt
@@ -2,5 +2,5 @@ package io.github.theapache64.perfsheet.model
 
 data class Node(
     val threadName: String,
-    val durationInMs: Double
+    val durationInMs: Long
 )
